+++ lib/datamappify/repository/query_method/update.rb
@@ -1,0 +1,8 @@
+module Datamappify
+  module Repository
+    module QueryMethod
+      class Update < Save
+      end
+    end
+  end
+end